extracted legend and axes to specific file
